pyproject.toml
[project]
name = "reposwarm-askbox"
version = "0.1.0"
description = "Docker-based ask agent for RepoSwarm — query your architecture docs with AI"
requires-python = ">=3.11"
license = "Apache-2.0"
dependencies = [
    "strands-agents>=0.1.0",
    "strands-agents-tools>=0.1.0",
    "claude-agent-sdk>=0.1.0",
    "boto3>=1.35.0",
    "anthropic>=0.40.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
]

[project.scripts]
askbox = "src.agent:main"
